requirements.txt
imageio
av
imageio-ffmpeg
pygifsicle
threadpoolctl
pillow
numpy
matplotlib
cython
quicktions
